requirements.txt
pynput
keyboard
pystray
Pillow
cx_Freeze
PyInstaller